+++ src/components/features/pages/my-info/EmergencyPersonalInfoSection/index.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { useModal } from "@hooks/common/useModal";
+// import { useModal } from "@hooks/common/useModal";
 import { useEmergencyPersonalInfo } from "@hooks/feature/info/useEmergencyPersonalInfo";
 import CancelIcon from "@icons/CancelIcon";
 import Spacing from "@shared/layout/Spacing";
 import Text from "@shared/ui/Text";
 import TextField from "@shared/ui/TextField";
-import WithdrawMemberModal from "@widgets/WithdrawMemberModal";
+// import WithdrawMemberModal from "@widgets/WithdrawMemberModal";
 
 export default function EmergencyPersonalInfoSection() {
   const {
@@ -15,7 +15,7 @@ export default function EmergencyPersonalInfoSection() {
     onClearEmergencyPersonalInfo,
   } = useEmergencyPersonalInfo();
 
-  const { closeModal, isOpen, openModal } = useModal();
+  // const { closeModal, isOpen, openModal } = useModal();
 
   return (
     <section>
@@ -83,7 +83,7 @@ export default function EmergencyPersonalInfoSection() {
         }
       />
       <Spacing size={8} />
-      <button onClick={openModal} className="flex justify-center w-full">
+      {/* <button onClick={openModal} className="flex justify-center w-full">
         <Text
           className="text-center"
           color="text-main-green"
@@ -91,10 +91,10 @@ export default function EmergencyPersonalInfoSection() {
         >
           회원 탈퇴
         </Text>
-      </button>
-      {isOpen && (
+      </button> */}
+      {/* {isOpen && (
         <WithdrawMemberModal closeModal={closeModal} openModal={openModal} />
-      )}
+      )} */}
     </section>
   );
 }
