+++ src/hooks/feature/bridge/useRouteCourseDetail/index.ts
